requirements.txt
requests
beautifulsoup4
langchain-openai
python-dotenv
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
